requirements.txt
av==14.2.1
msgpack==1.1.0
numpy==2.4.4
OpenEXR==3.2.4
OpenImageIO==3.0.6.0
pydub==0.25.1
PyOpenGL==3.1.10
PySide6==6.9.0
rawgpy==1.0.3
sounddevice==0.5.5
